Revert "Switch to `xorg-xproto` instead of the CDT"
This reverts commit 9f235617f9c50f96b6fddd5da6303c124da55464.

It seems using the `conda-forge` package for `xorg-xproto` doesn't quite
work as it appears it causes the CDT `libxxf86vm-devel` to miss a few
headers it expects to find. As a result, the build fails with the new
compilers. Reverting to the CDT in this case should fix the issue.

ref: https://circleci.com/gh/conda-forge/freeglut-feedstock/97